new contracts

+++ 2-lottery/contracts/Lottery.sol
@@ -10,31 +10,30 @@ contract Lottery {
 
     function enter() public payable {
         require(msg.value > 0.01 ether);
-        
+
         players.push(msg.sender);
     }
 
-    function random() private view returns(uint) {
+    function random() private view returns (uint256) {
         // sha3(block.difficulty, now, players);
-        return uint(keccak256(block.difficulty, now, players));
+        return uint256(keccak256(block.difficulty, now, players));
     }
 
     function pickWinner() public restricted {
         // require(msg.sender == manager);
 
-        uint index = random() % players.length;
+        uint256 index = random() % players.length;
         players[index].transfer(this.balance);
 
         players = new address[](0);
     }
-    
+
     modifier restricted() {
         require(msg.sender == manager);
         _;
     }
 
-    function getPlayers() public view returns(address[]) {
+    function getPlayers() public view returns (address[]) {
         return players;
     }
-    
 }
